requirements.txt
# Core Framework
safedeepagent>=0.1.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Networking & Traffic Analysis
scapy>=2.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=23.1.0
pyyaml>=6.0.0
psutil>=5.9.0

# CLI & UI
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0
